moved the finished checklist functionality into the
applicationform app, this makes sense as this app is focused on the
configuration of an application form
