+++ checklist/urls.py
@@ -22,6 +22,6 @@ urlpatterns = patterns('',
                     'success_name':'on_success',
                     }),                  
       url(r'^finish/(?P<ethics_application_id>\d+)/$',
-          view ='checklist.views.finished_checklist', 
+          view ='applicationform.views.finished_checklist', 
           name='finished_checklist'),
 )
